GEN AI/Multiagent research system/requirements.txt
langchain
langchain-google-genai
langchain-community
python-dotenv
gradio
beautifulsoup4
requests
ddgs
duckduckgo-search